all: 64-bit event numbers (eventstore >=4.0.0)
 - change protobuf definitions to use int64 instead
   of int32 in order to move along with eventstore
   protocol. Adjustments have been accordingly done
   with regards to `Long` vs. `Int` widening and
   signatures for `EventNumber`.

 - add test factor for akka in order to make slow
   CI runs more robust.

 - add `IdentifyClient` and `ClientIdentified` messages
   with corresponding marshalling and inspection/operations.

 - add new setting for client name.

 - update read me to reflect that ES client from now on
   only supports ES 4.0.0+.

 - adjust tests to account for ES client identification.

 - bump various dependencies.
 - downgrade spec2 as 3.8.7+ has concurrency issues.

 - fix `ProjectionsClientITest` code wrt. type mismatch.

+++ src/main/java/eventstore/j/EsConnection.java
@@ -377,7 +377,7 @@ public interface EsConnection {
   Closeable subscribeToStreamFrom(
       String stream,
       SubscriptionObserver<Event> observer,
-      Integer fromEventNumberExclusive,
+      Long fromEventNumberExclusive,
       boolean resolveLinkTos,
       UserCredentials credentials);
